Refactoring events Sockets
Déplacement des fonctions sockets dans un fichier dédié, ajout également d'un chargeur automatique des fonctions.
